requirements.txt
# ML / Modeling
xgboost==2.0.3
lightgbm==4.3.0
mlflow==2.9.2
joblib==1.3.2

# API & App
fastapi==0.110.0
uvicorn==0.27.1
streamlit==1.30.0

# Monitoring & Streaming
prometheus-client==0.19.0
kafka-python==2.0.2

# Utils
requests==2.31.0

# Testing
pytest==7.4.4
